chore: organize view structure
