+++ composeApp/src/main/kotlin/view/App.kt
@@ -2,19 +2,14 @@ package view
 
 import androidx.compose.animation.core.*
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.offset
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import cafe.adriel.voyager.navigator.CurrentScreen
 import cafe.adriel.voyager.navigator.Navigator
-import cafe.adriel.voyager.navigator.tab.TabNavigator
 import cafe.adriel.voyager.transitions.FadeTransition
-import org.jetbrains.compose.ui.tooling.preview.Preview
 import view.home.HomeScreen
 
 @Composable
-@Preview
 fun App() {
     MaterialTheme {
         Navigator(HomeScreen()) { navigator ->
