+++ composeApp/src/main/kotlin/view/workspace/graphView/GraphView.kt
@@ -1,0 +1,9 @@
+package view.workspace.graphView
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun GraphView(modifier: Modifier) {
+
+}
